Scheduling
Switched the mechanism to check on lock status from polling to
XL Release's scheduling mechanism. This ensures that if the server
is restarted while waiting on a lock it will still be waiting on
it correctly when the server comes back online. Additionally,
added better icons for the lock/unlock tasks.